Rename native_msg reader to seviri_native
