+++ satpy/etc/readers/seviri_native.yaml
@@ -1,13 +1,13 @@
 reader:
-  description: MSG Native Reader
-  name: native_msg
+  description: MSG Seviri Native Reader
+  name: seviri_native
   sensors: [seviri]
   default_channels: [HRV, IR_016, IR_039, IR_087, IR_097, IR_108, IR_120, IR_134, VIS006, VIS008, WV_062, WV_073]
   reader: !!python/name:satpy.readers.yaml_reader.FileYAMLReader
 
 file_types:
   native_msg:
-    file_reader: !!python/name:satpy.readers.native_msg.NativeMSGFileHandler ''
+    file_reader: !!python/name:satpy.readers.seviri_msg.NativeMSGFileHandler ''
     file_patterns: ['{satid:4s}-{instr:4s}-MSG{product_level:2d}-0100-NA-{processing_time1:%Y%m%d%H%M%S.%f}000Z-{order_id:s}.nat']
 
 datasets:
