+++ satpy/readers/seviri_native_hdr.py
